requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
django-mongoengine==0.3.3
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==3.2.2
mongoengine==0.20.0
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pymongo==3.10.1
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
